DOCS: 도커 파일 및 도커 컴포즈 파일 추가

+++ flask-sqlAlchemy/models/models.py
@@ -1,7 +1,7 @@
 from extensions import db
 
 class Member(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
+    id = db.Column(db.BigInteger, primary_key=True)
     name = db.Column(db.String(50))
     email = db.Column(db.String(100), unique=True, nullable=False)
